Adicionando o listener de mensagens privadas

+++ lib/src/main/kotlin/dev/kotlinautas/twitch4k/entity/User.kt
@@ -1,8 +1,8 @@
 package dev.kotlinautas.twitch4k.entity
 
-data class User (
-    val id: Int,
-    val displayName:String,
+data class User(
+    val id: Long,
+    val displayName: String,
     val isSubscriber: Boolean,
     val isModerator: Boolean,
     val isFirstMessage: Boolean,
